client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "UART Client for nRF54 connection time synchronization sample"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bleak>=0.22",
    "pyserial>=3.5",
]
